Replace BowlingGameTests with new Tests.
